@@ -11,6 +11,7 @@ import com.Test.RANKMFSELENIUM.HomepageTest;
 import com.Test.RANKMFSELENIUM.LoginPage;
 
 import RankmfBase.TestBase;
+import utility.Constant;
 
 import com.Test.RANKMFSELENIUM.SmartSwitchportfolio;
 
@@ -43,7 +44,7 @@ public class Basketsmartswitchportfolio  extends TestBase {
 		{
 			
 			LoginPage.clicklogintest();
-			LoginPage.smartswitchlogin();
+			LoginPage.smartswitchlogin(Constant.Username, Constant.password);
 			
 		}
 		@Test(groups= {"Regression"},dependsOnMethods= {"ClickonSigninPage"})
